move giou related unit test from test_losses.py to test_iou2d_calculator.py
